src/main/java/com/njustz/wechat_shopping/service/impl/ProductCategoryServiceImpl.java
package com.njustz.wechat_shopping.service.impl;

import com.njustz.wechat_shopping.entity.ProductCategory;
import com.njustz.wechat_shopping.repository.ProductCategoryRepository;
import com.njustz.wechat_shopping.service.ProductCategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * @author njustz
 * @date 2019/5/26
 */
@Service
public class ProductCategoryServiceImpl implements ProductCategoryService {

    @Autowired
    private ProductCategoryRepository productCategoryRepository;

    @Override
    public ProductCategory findOne(Integer productCategoryId) {
        return productCategoryRepository.findById(productCategoryId).get();
    }

    @Override
    public List<ProductCategory> findAll() {
        return productCategoryRepository.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryTypeList) {
        return productCategoryRepository.findByCategoryTypeIn(categoryTypeList);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return productCategoryRepository.save(productCategory);
    }
}
